Adding GuidedProjectile
This is the alternate attack of the TriangleShip. It also has a cooldown
value, i.e., its alternate attack only can be perfomed again after X
seconds.

It isn't complete developed, as we can set a target to the missile, so
it can adjust its trajectory to reach its aim.